refactor: 참가자의 상태를 나타내고 관리하는 State의 파일 분할 및 리팩터링
