+++ src/main/kotlin/blackjack/model/state/CompareImpl.kt
@@ -1,0 +1,8 @@
+package blackjack.model.state
+
+import blackjack.base.BaseHolder
+import blackjack.model.GameResult
+
+interface CompareImpl {
+    fun decideWinner(opponent: BaseHolder): GameResult
+}
